Rename + removed unused widgets
