+++ packages/smooth_app/lib/cards/data_cards/smooth_data_card.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:smooth_app/smooth_ui_library/animations/smooth_reveal_animation.dart';
+import 'package:smooth_app/generic_lib/animations/smooth_reveal_animation.dart';
 
 enum SmoothDataCardFormat { SQUARE, WIDE }
