+++ packages/smooth_app/lib/cards/product_cards/knowledge_panels/knowledge_panel_element_card.dart
@@ -6,7 +6,7 @@ import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_
 import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_group_card.dart';
 import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_table_card.dart';
 import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_world_map_card.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 
 class KnowledgePanelElementCard extends StatelessWidget {
   const KnowledgePanelElementCard({
