+++ packages/smooth_app/lib/cards/product_cards/knowledge_panels/knowledge_panel_group_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:openfoodfacts/model/KnowledgePanelElement.dart';
 import 'package:openfoodfacts/model/KnowledgePanels.dart';
 import 'package:smooth_app/cards/product_cards/knowledge_panels/knowledge_panel_card.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 
 class KnowledgePanelGroupCard extends StatelessWidget {
   const KnowledgePanelGroupCard({
