+++ packages/smooth_app/lib/cards/product_cards/question_card.dart
@@ -6,9 +6,9 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:smooth_app/cards/product_cards/product_image_carousel.dart';
 import 'package:smooth_app/cards/product_cards/product_title_card.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 import 'package:smooth_app/helpers/user_management_helper.dart';
 import 'package:smooth_app/pages/user_management/login_page.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
 
 class QuestionCard extends StatefulWidget {
   const QuestionCard({
