+++ packages/smooth_app/lib/cards/product_cards/smooth_product_card_found.dart
@@ -5,11 +5,11 @@ import 'package:openfoodfacts/model/Product.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_app/cards/data_cards/svg_icon_chip.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_product_image.dart';
 import 'package:smooth_app/helpers/product_cards_helper.dart';
 import 'package:smooth_app/helpers/product_compatibility_helper.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 import 'package:smooth_app/pages/product/new_product_page.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_product_image.dart';
 
 class SmoothProductCardFound extends StatelessWidget {
   const SmoothProductCardFound({
