+++ packages/smooth_app/lib/cards/product_cards/smooth_product_card_not_found.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/model/Product.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_simple_button.dart';
 import 'package:smooth_app/pages/product/new_product_page.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_simple_button.dart';
 
 class SmoothProductCardNotFound extends StatelessWidget {
   const SmoothProductCardNotFound({
