+++ packages/smooth_app/lib/generic_lib/animations/smooth_animated_collapse_arrow.dart
