+++ packages/smooth_app/lib/generic_lib/animations/smooth_reveal_animation.dart
