+++ packages/smooth_app/lib/generic_lib/dialogs/smooth_category_picker.dart
