+++ packages/smooth_app/lib/generic_lib/widgets/smooth_product_image.dart
