+++ packages/smooth_app/lib/generic_lib/widgets/smooth_toggle.dart
