+++ packages/smooth_app/lib/helpers/ui_helpers.dart
