+++ packages/smooth_app/lib/pages/abstract_user_preferences.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 
 /// Abstraction of a collapsed/expanded display for the preferences page.
 abstract class AbstractUserPreferences {
