+++ packages/smooth_app/lib/pages/onboarding/next_button.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 import 'package:smooth_app/pages/onboarding/onboarding_flow_navigator.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 
 /// Next button showed at the bottom of the onboarding flow.
