+++ packages/smooth_app/lib/pages/onboarding/scan_example.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 import 'package:smooth_app/pages/onboarding/next_button.dart';
 import 'package:smooth_app/pages/onboarding/onboarding_flow_navigator.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
 
 /// Example explanation on how to scan a product.
 class ScanExample extends StatelessWidget {
