+++ packages/smooth_app/lib/pages/product/common/product_dialog_helper.dart
@@ -5,8 +5,8 @@ import 'package:smooth_app/data_models/fetched_product.dart';
 import 'package:smooth_app/database/barcode_product_query.dart';
 import 'package:smooth_app/database/dao_product.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_app/smooth_ui_library/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_simple_button.dart';
+import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 
 /// Dialog helper for product barcode search
 class ProductDialogHelper {
