+++ packages/smooth_app/lib/pages/product/common/product_list_dialog_helper.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/database/dao_product_list.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_app/smooth_ui_library/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_simple_button.dart';
+import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
 
 class ProductListDialogHelper {
   @visibleForTesting
