+++ packages/smooth_app/lib/pages/product/nutrition_page_loaded.dart
@@ -9,9 +9,9 @@ import 'package:openfoodfacts/model/OrderedNutrients.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:openfoodfacts/utils/UnitHelper.dart';
 import 'package:smooth_app/database/product_query.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_app/smooth_ui_library/dialogs/smooth_alert_dialog.dart';
-import 'package:smooth_app/smooth_ui_library/util/ui_helpers.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_simple_button.dart';
+import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_app/helpers/ui_helpers.dart';
 
 /// Actual nutrition page, with data already loaded.
 class NutritionPageLoaded extends StatefulWidget {
