+++ packages/smooth_app/lib/pages/user_management/forgot_password_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_card.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_text_form_field.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_text_form_field.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
 
 class ForgotPasswordPage extends StatefulWidget {
