+++ packages/smooth_app/lib/pages/user_management/login_page.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:provider/provider.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_text_form_field.dart';
 import 'package:smooth_app/helpers/user_management_helper.dart';
 import 'package:smooth_app/pages/user_management/forgot_password_page.dart';
 import 'package:smooth_app/pages/user_management/sign_up_page.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_card.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_text_form_field.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
 
 // TODO(M123-dev): Handle colors better
