+++ packages/smooth_app/lib/pages/user_preferences_settings.dart
@@ -3,13 +3,13 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:smooth_app/data_models/user_preferences.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_main_button.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_simple_button.dart';
+import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_list_tile.dart';
+import 'package:smooth_app/generic_lib/widgets/smooth_toggle.dart';
 import 'package:smooth_app/helpers/launch_url_helper.dart';
 import 'package:smooth_app/pages/abstract_user_preferences.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_main_button.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_simple_button.dart';
-import 'package:smooth_app/smooth_ui_library/dialogs/smooth_alert_dialog.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_list_tile.dart';
-import 'package:smooth_app/smooth_ui_library/widgets/smooth_toggle.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 import 'package:smooth_app/themes/theme_provider.dart';
 import 'package:smooth_app/views/bottom_sheet_views/user_contribution_view.dart';
