--- packages/smooth_app/lib/smooth_ui_library/widgets/models/single_boolean_model.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class SingleBooleanModel extends ChangeNotifier {
-  bool isActive = false;
-
-  void setActive() {
-    isActive = true;
-    notifyListeners();
-  }
-
-  void setInactive() {
-    isActive = false;
-    notifyListeners();
-  }
-}
