+++ packages/smooth_app/lib/views/bottom_sheet_views/group_query_filter_view.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
-import 'package:smooth_app/smooth_ui_library/buttons/smooth_main_button.dart';
+import 'package:smooth_app/generic_lib/buttons/smooth_main_button.dart';
 
 class GroupQueryFilterView extends StatelessWidget {
   const GroupQueryFilterView({
